Move `TransactionContext` into submodule
